requirements.txt
flask
flask-cors
google-generativeai
PyPDF2
python-docx
python-dotenv
